tools/skill_builder/requirements.txt
PyGithub==2.1.1
gitpython==3.1.43
requests==2.31.0
python-dateutil==2.8.2
diskcache==5.6.3
